ref: restructuring initialiseSession, unnesting configure plugin
