--- runner/src/server/plugins/initialiseSession/configurePlugin.ts
@@ -1,10 +1,0 @@
-import { initialiseSession } from "server/plugins/initialiseSession/initialiseSession";
-
-export function configureInitialiseSessionPlugin(options: {
-  safelist: string[];
-}) {
-  return {
-    plugin: initialiseSession,
-    options,
-  };
-}
