--- runner/src/server/plugins/initialiseSession/index.ts
@@ -1,1 +1,0 @@
-export { initialiseSession } from "./initialiseSession";
